Refactor: 마이페이지와 프로필 페이지를 하나로 합침(사용자 본인 프로필일땐 userId가 0)
